components-architecture
