+++ package.json
@@ -15,7 +15,6 @@
         "axios": "^0.27.2",
         "chart.js": "^3.8.2",
         "formik": "^2.2.9",
-        "jwt-decode": "^3.1.2",
         "modern-normalize": "^1.1.0",
         "prop-types": "^15.8.1",
         "rc-progress": "^3.4.0",
